+++ models/experimental/detr3d/tests/pcc/test_ttnn_pointnet_samodule_votes.py
@@ -6,7 +6,7 @@ import torch
 import pytest
 import ttnn
 from models.experimental.detr3d.reference.detr3d_model import PointnetSAModuleVotes
-from models.experimental.detr3d.ttnn.ttnn_pointnet_samodule_votes import TtnnPointnetSAModuleVotes
+from models.experimental.detr3d.ttnn.pointnet_samodule_votes import TtnnPointnetSAModuleVotes
 from tests.ttnn.utils_for_testing import comp_pcc
 from ttnn.model_preprocessing import preprocess_model_parameters
 from models.experimental.detr3d.ttnn.custom_preprocessing import create_custom_mesh_preprocessor
