+++ models/experimental/detr3d/ttnn/multihead_attention.py
@@ -4,12 +4,10 @@
 
 import ttnn
 import math
-from models.common.lightweightmodule import LightweightModule
 
 
-class TtnnMultiheadAttention(LightweightModule):
-    def __init__(self, d_model, nhead, device, parameters=None):
-        super().__init__()
+class TTNNMultiheadAttention:
+    def __init__(self, d_model, nhead, device):
         self.d_model = d_model
         self.nhead = nhead
         self.head_dim = d_model // nhead
@@ -25,30 +23,11 @@ class TtnnMultiheadAttention(LightweightModule):
         self.out_weight = None
         self.out_bias = None
 
-        # Load parameters if provided
-        if parameters is not None:
-            self.load_parameters(parameters)
-
-    def load_parameters(self, parameters):
-        """Load preprocessed attention parameters"""
-        self.q_weight = parameters.get("q_weight")
-        self.k_weight = parameters.get("k_weight")
-        self.v_weight = parameters.get("v_weight")
-        self.q_bias = parameters.get("q_bias")
-        self.k_bias = parameters.get("k_bias")
-        self.v_bias = parameters.get("v_bias")
-        self.out_weight = parameters.get("out_weight")
-        self.out_bias = parameters.get("out_bias")
-
-    def forward(self, query, key, value, attn_mask=None):
+    def __call__(self, query, key, value, attn_mask=None):
         # Apply linear projections separately to avoid concat issues
-        q = ttnn.to_memory_config(query, ttnn.L1_MEMORY_CONFIG)
-        k = ttnn.to_memory_config(key, ttnn.L1_MEMORY_CONFIG)
-        v = ttnn.to_memory_config(value, ttnn.L1_MEMORY_CONFIG)
-
-        q = ttnn.linear(q, self.q_weight, bias=self.q_bias, memory_config=ttnn.L1_MEMORY_CONFIG)
-        k = ttnn.linear(k, self.k_weight, bias=self.k_bias, memory_config=ttnn.L1_MEMORY_CONFIG)
-        v = ttnn.linear(v, self.v_weight, bias=self.v_bias, memory_config=ttnn.L1_MEMORY_CONFIG)
+        q = ttnn.linear(query, self.q_weight, bias=self.q_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        k = ttnn.linear(key, self.k_weight, bias=self.k_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        v = ttnn.linear(value, self.v_weight, bias=self.v_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
 
         # Get dimensions for reshaping
         batch_size = q.shape[0]
@@ -58,13 +37,13 @@ class TtnnMultiheadAttention(LightweightModule):
         # Reshape each tensor separately for multi-head attention
         # [batch, seq_len, d_model] -> [batch, seq_len, num_heads, head_dim] -> [batch, num_heads, seq_len, head_dim]
         q = ttnn.reshape(q, (batch_size, q_seq_len, self.nhead, self.head_dim))
-        q = ttnn.permute(q, (0, 2, 1, 3), memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        q = ttnn.permute(q, (0, 2, 1, 3))
 
         k = ttnn.reshape(k, (batch_size, k_seq_len, self.nhead, self.head_dim))
-        k = ttnn.permute(k, (0, 2, 1, 3), memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        k = ttnn.permute(k, (0, 2, 1, 3))
 
         v = ttnn.reshape(v, (batch_size, k_seq_len, self.nhead, self.head_dim))
-        v = ttnn.permute(v, (0, 2, 1, 3), memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        v = ttnn.permute(v, (0, 2, 1, 3))
 
         # Use SDPA for attention computation
         context = ttnn.transformer.scaled_dot_product_attention(
@@ -76,15 +55,11 @@ class TtnnMultiheadAttention(LightweightModule):
             attn_mask=attn_mask,
             memory_config=ttnn.DRAM_MEMORY_CONFIG,
         )
-        ttnn.deallocate(q)
-        ttnn.deallocate(k)
-        ttnn.deallocate(v)
 
         # Concatenate heads back to original format
-        context = ttnn.transformer.concatenate_heads(context, memory_config=ttnn.L1_MEMORY_CONFIG)
+        context = ttnn.transformer.concatenate_heads(context, memory_config=ttnn.DRAM_MEMORY_CONFIG)
 
         # Output projection
-        output = ttnn.linear(context, self.out_weight, bias=self.out_bias, memory_config=ttnn.L1_MEMORY_CONFIG)
-        ttnn.deallocate(context)
+        output = ttnn.linear(context, self.out_weight, bias=self.out_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
 
         return output
