+++ models/experimental/detr3d/ttnn/shared_mlp.py
@@ -3,50 +3,39 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.common.lightweightmodule import LightweightModule
-from models.experimental.detr3d.ttnn.utils import TtnnConv2D
+from models.experimental.detr3d.ttnn.common import TtnnConv2D
 
 
-class TtnnSharedMLP(LightweightModule):
-    def __init__(self, parameters, device):
-        super().__init__()
+class TtnnSharedMLP:
+    def __init__(self, module, parameters, device):
         self.device = device
         self.parameters = parameters
-        shard_layout = ttnn.TensorMemoryLayout.HEIGHT_SHARDED
         self.conv1 = TtnnConv2D(
-            parameters.conv_args.layer0.conv,
+            module.layer0.conv,
             parameters.layer0.conv,
             device,
             activation=ttnn.UnaryWithParam(ttnn.UnaryOpType.RELU),
-            is_dealloc_act=True,
+            is_dealloc_act=False,
             return_dims=True,
-            shard_layout=shard_layout,
-            math_fidelity=ttnn.MathFidelity.HiFi2,
         )
         self.conv2 = TtnnConv2D(
-            parameters.conv_args.layer1.conv,
+            module.layer1.conv,
             parameters.layer1.conv,
             device,
             activation=ttnn.UnaryWithParam(ttnn.UnaryOpType.RELU),
             is_dealloc_act=True,
             return_dims=True,
-            shard_layout=shard_layout,
-            memory_config=ttnn.DRAM_MEMORY_CONFIG,
-            math_fidelity=ttnn.MathFidelity.HiFi2,
         )
         self.conv3 = TtnnConv2D(
-            parameters.conv_args.layer2.conv,
+            module.layer2.conv,
             parameters.layer2.conv,
             device,
             activation=ttnn.UnaryWithParam(ttnn.UnaryOpType.RELU),
             is_dealloc_act=True,
             return_dims=True,
-            shard_layout=shard_layout,
-            memory_config=ttnn.DRAM_MEMORY_CONFIG,
-            math_fidelity=ttnn.MathFidelity.HiFi2,
         )
 
-    def forward(self, features):
+    def __call__(self, features):
         shape = features.shape
         conv1, shape = self.conv1(features, shape)
         conv2, shape = self.conv2(conv1, shape)
